requirements.txt
aiofiles==0.7.0
aiosqlite==0.16.1
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
databases==0.4.3
fastapi==0.65.1
h11==0.12.0
html5lib==1.1
idna==2.10
iso8601==0.1.14
Jinja2==3.0.1
MarkupSafe==2.0.1
pydantic==1.8.2
pypika-tortoise==0.1.1
python-multipart==0.0.5
pytz==2021.1
requests==2.25.1
six==1.16.0
soupsieve==2.2.1
SQLAlchemy==1.3.24
starlette==0.14.2
tortoise-orm==0.17.3
typing-extensions==3.10.0.0
urllib3==1.26.4
uvicorn==0.13.4
webencodings==0.5.1
